requirements.txt
# SakhiSetu - Women Government Scheme Navigator
# Python 3.10+ (3.14 supported with pydantic>=2.12.5)
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy>=2.0.25

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.3
playwright>=1.40.0

# Validation & Config (pydantic>=2.12.5 required for Python 3.14)
pydantic>=2.12.5
pydantic-settings>=2.6.0
python-dotenv>=1.0.1

# Utilities
python-multipart>=0.0.9

# Translation (multilingual)
deep-translator>=1.11.0

# RAG Chatbot (TF-IDF retrieval + Gemini)
google-generativeai>=0.8.0
scikit-learn>=1.3.0

# Scheduled tasks (auto-scrape)
apscheduler>=3.10.0
